requirements.txt
# Core trading and data libraries
backtrader==1.9.78.123
yfinance==0.2.18
alpaca-trade-api==3.1.1
pandas==2.0.3
numpy==1.24.3

# Configuration and utilities
PyYAML==6.0.1
python-dotenv==1.0.0

# Visualization and analysis
matplotlib==3.7.2

# Development and testing
pytest==7.4.0
